backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
qdrant-client
fastembed
biopython
chonkie[model2vec]
numpy
python-dotenv
requests
<<<<<<< HEAD
tensorflow
rdkit
openai
=======
openai
tensorflow
>>>>>>> d2ea12fb70eefdfe3e74f4e0253b9bb073cfc306
